管理系统bug修复提交

+++ biz-feature-admin/biz-feature-admin-srv/src/main/java/com/ly/ttd/biz/feature/admin/mybatis/entity/SequenceEntity.java
@@ -32,5 +32,5 @@ public class SequenceEntity {
     /**
      * 序列值
      */
-    private Long val;
+    private Integer val;
 }
